Fix /news endpoint: cast  to integer for PostgreSQL date subtraction
asyncpg encodes Python int as int8 (bigint), but PostgreSQL only has
date - integer (int4) operator, not date - bigint.

+++ stock-ai-agent/agents/news_agent.py
@@ -114,7 +114,7 @@ async def get_recent_news(stock_code: str, days: int = 5) -> list[dict]:
     return await fetch_all("""
         SELECT title, source, sentiment, news_date
         FROM news_cache
-        WHERE stock_code = $1 AND news_date >= CURRENT_DATE - $2
+        WHERE stock_code = $1 AND news_date >= CURRENT_DATE - $2::integer
         ORDER BY news_date DESC, id DESC
         LIMIT 10
     """, stock_code, days)
